[Bronze V] Title: 이상한 기호, Time: 52 ms, Memory: 5748 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -5,11 +5,11 @@ class Program
 {
     static void Main()
     {
-        var nums = Console.ReadLine().Split().Select(int.Parse).ToArray();
+        var nums = Console.ReadLine().Split().Select(long.Parse).ToArray();
         var(A, B) = (nums[0], nums[1]);
         Console.WriteLine(Atmark(A, B));
     }
-    static int Atmark(int a, int b)
+    static long Atmark(long a, long b)
     {
         return (a + b)*(a - b);
     }
